Move humanizeCoercionStrategy
