+++ frontend/src/metabase/metadata/components/CoercionStrategyPicker/utils.ts
